fix: handle on-ledger requests with malformed metadata
